Remove mouse reporting so terminals keep native selection
Mouse reporting (SGR 1000/1002/1006) made the terminal hand every
click and drag to the app, which is why drag-select and copy did not
work over the TUI. Rip it out:

- mouse.ts (dispatcher, hit rects, SGR parsing, MOUSE_ON/OFF) deleted;
  Selectable drops its rect registration and App/Chat their click and
  wheel listeners.
- The stdin tap survives as stdinTap.ts without mouse parsing — it is
  still what silences Ink while a clack prompt borrows the terminal.
- Keyboard paths are unchanged: every clickable row was already
  reachable with arrows and Enter.
